refactor(fhir): TEFCA QHIN Phase E-2 — _shared/fhir relocation + bulk cleanup + signed-URL mode
Closes the Phase E-1 follow-ups in one PR.

E-3: relocate cross-function FHIR helpers to supabase/functions/_shared/fhir/
- Moved verbatim:
    bearer-auth.ts   (bearer/legacy auth resolver, scope checks)
    audit.ts         (logTEFCAAccess + verifyPatientConsent)
    mappers.ts       (14 row -> FHIR R4 mappers)
    registry.ts      (RESOURCE_REGISTRY)
- Split tefca-ias/shared.ts: FHIR-domain pieces (SYSTEM_IDENTIFIERS,
  US_CORE, VITAL_LOINC_CODES, SDOH_LOINC_CODES, ExchangePurpose,
  TEFCAContext, createBundle, createOperationOutcome, VALID_EXCHANGE_PURPOSES)
  moved to _shared/fhir/codes.ts. Endpoint-specific HTTP helpers
  (corsHeaders, fhirJsonHeaders, fhirJsonResponse, errorResponse) stay in
  tefca-ias/shared.ts and re-export the FHIR domain symbols for
  backward-compatible imports.
- Updated consumers: tefca-ias/{index,history,capability}.ts and
  tefca-bulk/{index,processor}.ts now import from ../_shared/fhir/* —
  the canonical Supabase pattern. Cross-function imports (..  /tefca-ias/*)
  are gone.

E-2.1: pg_cron cleanup of expired bulk_export_jobs
- New migration 20260503050000_add_bulk_export_cleanup_cron.sql:
    CREATE EXTENSION IF NOT EXISTS pg_cron;
    cleanup_expired_bulk_exports() (SECURITY DEFINER, search_path scoped):
      1. flips stuck non-terminal jobs to status='failed' so /bulk-status
         reports a useful state past expires_at
      2. drops storage.objects matching <jobId>/* in the fhir-bulk bucket
      3. cascades through bulk_export_files and bulk_export_jobs rows
    Schedule: every hour at minute 7 ('7 * * * *'), idempotent (unschedule
    + re-schedule). tefca_access_logs are intentionally NOT touched here —
    those keep the full TEFCA RCE 6-year retention regardless of bulk
    cleanup.

E-2.2: signed-URL redirect mode for large transfers
- GET /bulk-files/{jobId}/{file}?redirect=signed now 302s to a Supabase
  Storage signed URL with the existing SIGNED_URL_TTL_SECONDS lifetime,
  for clients that prefer to pull directly from Storage. Default behavior
  (no query param) still streams through the function for per-file audit.
- Removed the dead handleSignedRedirect stub that was reserved for this
  in Phase E-1.

E-2.3: Group/{id}/$export enforcement
- New ensureGroupAccessible(supabase, groupId) precheck. mBHR doesn't yet
  have a FHIR Group resource (no groups / group_membership tables), so
  the function:
    - tries SELECT id FROM groups WHERE id=$1
    - on error (e.g. relation does not exist) returns 501 not-supported
      with a pointer to the registry of supported flows
- Wired into handleKickoff before any job row is created, so a partner
  hitting /Group/<id>/$export gets a clean 501 instead of an opaque
  failure. JSDoc on ensureGroupAccessible spells out exactly what the
  Phase E-3 implementation should do once the Group resource lands
  (RLS-backed membership check, roster resolution into the processor).

Verification
- npx eslint --max-warnings=0 clean across _shared/fhir/, tefca-ias/, tefca-bulk/
- src tests still pass
